pyproject.toml
[project]
name = "visualcentricbenchmark"
version = "0.1.0"
description = "Diagram agent: textbook math diagrams via TikZ, Matplotlib, and Graphviz"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "graphviz>=0.21",
    "matplotlib>=3.10.8",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = false
